==> Scripts/03 Automaticity.Rmd <==
---
title: "Explore ExIm study FB"
author: "ddd"
date: "9/3/2020"
output: pdf_document
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "Markdown PDFs") })
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = F)
library(dplyr)
library(tidyr)
library(ggplot2)
library(ggthemes)
theme_set(theme_tufte())
options(ggplot2.continuous.colour="plasma")
options(ggplot2.continuous.fill = "plasma")

knitr::opts_knit$set(root.dir = '/Users/diagdavenport/Desktop/Synced Research/Ludwig/FB/Lockdown')

colors <- c("#E1BE6A", "#40B0A6")

```

```{r import data}
 
exim.raw <- read.csv('Data/Experimental Lab Data/Automaticity/ExIm data.csv')

```

```{r cleaning and organizing}
 
exim.raw$access_pymk <- ifelse(exim.raw$condition1 == "pymk", exim.raw$access1, exim.raw$access2)
exim.raw$access_nf <- ifelse(exim.raw$condition1 == "like", exim.raw$access1, exim.raw$access2)

exim.raw$cogeffort_pymk <- ifelse(exim.raw$condition1 == "pymk", exim.raw$cog_effort1, exim.raw$cog_effort2)
exim.raw$cogeffort_nf <- ifelse(exim.raw$condition1 == "like", exim.raw$cog_effort1, exim.raw$cog_effort2)

exim.raw$influence_pymk <- ifelse(exim.raw$condition1 == "pymk", exim.raw$influence1, exim.raw$influence2)
exim.raw$influence_nf <- ifelse(exim.raw$condition1 == "like", exim.raw$influence1, exim.raw$influence2)

exim.raw$gut_pymk <- ifelse(exim.raw$condition1 == "pymk", exim.raw$gut1, exim.raw$gut2)
exim.raw$gut_nf <- ifelse(exim.raw$condition1 == "like", exim.raw$gut1, exim.raw$gut2)

exim.raw$careful_pymk <- ifelse(exim.raw$condition1 == "pymk", exim.raw$careful1, exim.raw$careful2)
exim.raw$careful_nf <- ifelse(exim.raw$condition1 == "like", exim.raw$careful1, exim.raw$careful2)

exim.raw$awareness_pymk <- ifelse(exim.raw$condition1 == "pymk", exim.raw$awareness1, exim.raw$awareness2)
exim.raw$awareness_nf <- ifelse(exim.raw$condition1 == "like", exim.raw$awareness1, exim.raw$awareness2)

exim.raw$intention_pymk <- ifelse(exim.raw$condition1 == "pymk", exim.raw$intention1, exim.raw$intention2)
exim.raw$intention_nf <- ifelse(exim.raw$condition1 == "like", exim.raw$intention1, exim.raw$intention2)

exim.raw$control_pymk <- ifelse(exim.raw$condition1 == "pymk", exim.raw$controllability1, exim.raw$controllability2)
exim.raw$control_nf <- ifelse(exim.raw$condition1 == "like", exim.raw$controllability1, exim.raw$controllability2)

exim.raw$efficiency_pymk <- ifelse(exim.raw$condition1 == "pymk", exim.raw$efficiency1, exim.raw$efficiency2)
exim.raw$efficiency_nf <- ifelse(exim.raw$condition1 == "like", exim.raw$efficiency1, exim.raw$efficiency2)

exim.raw$speed_pymk <- ifelse(exim.raw$condition1 == "pymk", exim.raw$speed1, exim.raw$speed2)
exim.raw$speed_nf <- ifelse(exim.raw$condition1 == "like", exim.raw$speed1, exim.raw$speed2)


 higher.is.more.automatic <- c("access", "careful", "awareness", "control", "efficiency")
 
 higher.is.less.automatic <-c("cogeffort", "influence", "gut", "intention")

 for (root in higher.is.less.automatic) {
  exim.raw[[paste0(root, "_nf")]] <- max(exim.raw[[paste0(root, "_nf")]], na.rm = T) + 1 - exim.raw[[paste0(root, "_nf")]]
  exim.raw[[paste0(root, "_pymk")]] <- max(exim.raw[[paste0(root, "_pymk")]], na.rm = T) + 1 - exim.raw[[paste0(root, "_pymk")]]
 }
 
```

```{r graphing}

choice.long <- exim.raw %>% dplyr::select(c(ends_with("_nf"), ends_with("_pymk"))) %>% 
  pivot_longer(everything(),
   names_pattern = "(.*)_(.*)",
   names_to = c(".value", ".set"),
   values_to = "count"
 )

head(choice.long)

"new_?(.*)_(.)(.*)"

qs <- c("access",
        "cogeffort",
        "influence",
        "gut",
        "careful",
        "awareness",
        "intention",
        "control",
        "efficiency",
        "speed")

choice.long2 <- choice.long %>% 
  pivot_longer(cols = all_of(qs),
               names_to = "cat",
               values_to = "val") %>% dplyr::rename(target = ".set")


choice.long2$val <- ifelse(choice.long2$cat == "gut",
                 choice.long2$val - 1,
                 choice.long2$val)

choice.long2$val <- ifelse(choice.long2$cat == "careful",
                 choice.long2$val - 1,
                 choice.long2$val)

temp.l.dat <- choice.long2 %>% group_by(cat)  %>% dplyr::summarise(mean.val = mean(val, na.rm = T), sd = sd(val, na.rm = T), n= n()) 

l.dat <- choice.long2 %>% filter(cat == "speed") %>% group_by(cat, target)  %>% dplyr::summarise(mean.val = mean(val, na.rm = T), sd = sd(val, na.rm = T), n= n())

l.dat$ci <- (l.dat$sd/sqrt(l.dat$n))*1.96

 ggplot(l.dat, aes(x = cat, y = mean.val, fill = target)) +
  geom_col(position = "dodge") +
   geom_errorbar(aes(ymin=mean.val-ci, ymax=mean.val+ci),
                 width=.2,                    # Width of the error bars
                  position=position_dodge(.9))  + theme(axis.text.x = element_text(angle = 45, hjust = 1))
 
 dev.copy(jpeg,'Output/Graphs/Experiments/Automaticity/speed bar chart.jpg')
dev.off()

m.long <- merge(choice.long2, temp.l.dat, by = c("cat"))
m.long$z <- (m.long$val - m.long$mean.val)/m.long$sd
 
 label.df <- data.frame(cat = c("access",
                                "awareness",
                                "cogeffort", 
                                "control", 
                                "efficiency",
                                "influence", 
                                "intention",
                                "simple composite"),
                       mean.val = c(2.9, 2.8, 4, 2.4, 4.4, 4.4, 4, 3.6),
                       target = "nf")
 
automaticity.labels <- c(
                                "(Intention) Inability to resist considering",
                                "(Access) Difficult to explain why",
                                "(Unwareness) Unaware what affected decision", 
                                "(Control) Inability to resist action urge",
                                "(Influence) Infleunced by situation",  
                                "(Efficiency) Low cognitive effort", 
                                "(Cog Effort) Lack of forethought",
                                "Simple Composite Score")

automaticity.labels.full <- c(automaticity.labels[1:7],
                              "(Gut) Based on gut reaction",
                              "(Careful) Not a careful consideration",
                              "PCA (First component, 28% PVE)",
                              "Composite")

l.dat1 <- choice.long2 %>% filter(cat != "speed") %>% group_by(cat, target)  %>% dplyr::summarise(mean.val = mean(val, na.rm = T), sd = sd(val, na.rm = T), n= n()) 

l.dat1$ci <- (l.dat1$sd/sqrt(l.dat1$n))*1.96

l.dat2 <- choice.long2 %>% filter(cat != "speed" & cat != "gut" & cat != "careful") %>% group_by(target)  %>% dplyr::summarise(mean.val = mean(val, na.rm = T), sd = sd(val, na.rm = T), n= n()) 
l.dat2$cat <- "simple composite"
l.dat2$ci <- (l.dat2$sd/sqrt(l.dat2$n))*1.96

pca <- prcomp(choice.long[!is.na(choice.long$access),
                             c("access", "cogeffort", "influence", "awareness", "intention", "control", "efficiency")],
                 center = TRUE, scale. = TRUE)

pca.with.label <- cbind(choice.long[!is.na(choice.long$access), ".set"], pca$x)

l.dat3 <- pca.with.label %>% group_by(.set) %>% dplyr::summarise(mean.val = mean(PC1, na.rm = T), sd = sd(PC1, na.rm = T), n= n()) %>% mutate(target = .set)
l.dat3$cat <- "First Principal Component (28% PVE)"
l.dat3$ci <- (l.dat2$sd/sqrt(l.dat3$n))*1.96

l.dat <- rbind(as.data.frame(l.dat1), as.data.frame(l.dat2))

# sort the bars by effect size
l.dat$cat <- factor(l.dat$cat,levels = c("intention", "access", "awareness", "control", "influence", "efficiency", "cogeffort", "gut", "careful", "simple composite"))

 ggplot(l.dat %>% filter(cat != "careful" & cat != "gut"), aes(x = cat, y = mean.val, fill = target)) +
  geom_col(position = "dodge", aes(fill = target)) +
   geom_errorbar(aes(ymin=mean.val-ci, ymax=mean.val+ci),
                 width=.2,                    # Width of the error bars
                  position=position_dodge(.9)) + xlab("Measure")  +
   ylab("Mean Rescaled Automaticity Rating\n(Higher ratings = more automatic)") +
   labs(title = "Automaticity Ratings by Target",
        subtitle = "All dimensions of automaticity indicate that NF behavior is more automatic \nthan PYMK behavior.") +
   geom_text(data = label.df,
             label = c("NS", "**", "***", "***", "***", "***", "NS", "***"))  +
   scale_x_discrete(labels= automaticity.labels) + coord_flip() +
   scale_fill_manual(values=colors,
                       name="Target",
                       breaks=c("nf", "pymk"),
                       labels=c("Newsfeed", "PYMK"))
  
dev.copy(jpeg,'Output/Graphs/Experiments/Automaticity/composite score and components.jpg')
dev.off()
 
# pwr.t.test(d = .5, sig.level = .05, power = .8, type = c("paired"))

t <- l.dat %>% dplyr::select(cat, target, mean.val) %>% pivot_wider(names_from = target, values_from = mean.val)

ggplot(t, aes(x=nf, y=pymk, label = cat)) +
  geom_point(size=2, shape=23) + xlim(c(0,4.2)) + ylim(c(0,4.2)) + geom_text() + geom_abline(intercept = 0, slope = 1, linetype = "dotted")

dev.copy(jpeg,'Output/Graphs/Experiments/Automaticity/45 degree scatter.jpg')
dev.off()

cohen.data <- merge(t, temp.l.dat)

cohen.data$cohen.d <- with(cohen.data, (nf - pymk)/sd)

last.row <- data.frame(cat = "PCA",
                       "nf" = l.dat3[1,2],
                       "pymk" = l.dat3[2,2],
                       "mean.val" = 0,
                       sd = sd(pca.with.label$PC1),
                       n = 688,
                       cohen.d = abs(l.dat3[1,2] - l.dat3[2,2])/sd(pca.with.label$PC1))
colnames(last.row) <- names(cohen.data)

cohen.data <- rbind(cohen.data, last.row)


all.vals <- choice.long2 %>% filter(cat != "speed" & cat != "gut" & cat != "careful")

last.row <- data.frame(cat = "composite",
                       "nf" = mean(all.vals[all.vals$target == "nf",]$val, na.rm = T),
                       "pymk" = mean(all.vals[all.vals$target == "pymk",]$val, na.rm = T),
                       "mean.val" = NA,
                       sd = sd(all.vals$val, na.rm = T),
                       n = 688,
                       cohen.d = 777)
colnames(last.row) <- names(cohen.data)

cohen.data <- rbind(cohen.data, last.row)

cohen.data$cohen.d <- with(cohen.data, (nf - pymk)/sd)

cohen.data$var <- with(cohen.data, (n/(n/2)^2 + (cohen.d^2)/(2*n)))
  
cohen.data$ci <- with(cohen.data, sqrt(var)*1.96)
pd <- position_dodge(0.01)

 ggplot(cohen.data, aes(x = cat, y = abs(cohen.d))) +
   geom_errorbar(aes(ymin=abs(cohen.d)-ci, ymax=abs(cohen.d)+ci), colour="black", width=.1, position=pd) + ylim(c(-.8, .8)) +
  geom_point(position = "dodge") + geom_hline(yintercept = c(.2,.5,-.2, -.5), linetype = "dotted") + coord_flip() + geom_hline(yintercept = 0, linetype = "solid") + scale_x_discrete(labels= automaticity.labels.full) + ylab("Absolute Effect Size") + xlab("Measure")
 
 dev.copy(jpeg,'Output/Graphs/Experiments/Automaticity/standardized effect sizes.jpg')
dev.off()

  ggplot(cohen.data, aes(x = cat, y = abs(cohen.d))) +
  geom_point(position = "dodge") + geom_hline(yintercept = c(.2,.5,-.2, -.5), linetype = "dotted") + ylim(c(-.8, .8)) +
   geom_errorbar(aes(ymin=abs(cohen.d)-ci, ymax=abs(cohen.d)+ci), colour="black",
                 width=.2,                    # Width of the error bars
                  position=position_dodge(.9))
  


```


```{r continuous measures - excess mass}

label_names <- list(
  'access'="(Access) Difficult to explain why",
  'awareness'="(Unwareness) Unaware what \naffected decision",
  'intention'="(Intention) Inability to resist considering"
)

fun_labeller <- function(variable,value){
  return(label_names[value])
}

dat <- m.long %>% group_by(val, target, cat) %>% dplyr::summarise(n= n()) %>%
  pivot_wider(names_from = target, values_from = n, values_fill = 0)

dat <- dat %>% group_by(cat) %>% 
       mutate(perNF=nf/sum(nf)) %>% 
       ungroup

dat <- dat %>% group_by(cat) %>% 
       mutate(perPYMK=pymk/sum(pymk)) %>% 
       ungroup

dat$excess <- (dat$perNF - dat$perPYMK)/dat$perPYMK

l.dat <- m.long %>% filter(cat != "speed") %>% group_by(cat, target)  %>% dplyr::summarise(mean.z = mean(z, na.rm = T), sd = sd(z, na.rm = T), n= n())

l.dat$ci <- (l.dat$sd/sqrt(l.dat$n))*1.96

 ggplot(choice.long2, aes(x = val, fill = target)) +
  geom_histogram(aes(y=..density..), bins=7, alpha=0.5, position="nudge") +
   theme(axis.text.x = element_text(angle = 45, hjust = 1)) + xlab("Measure") +
   ylab("Response Density") + xlab("Rating (Higher ratings = more automatic)") + labs(title = "Automaticity Ratings by Target") + facet_wrap(~ cat, scales = "free")
 
  dev.copy(jpeg,'Output/Graphs/Experiments/Automaticity/histogram for each measure.jpg')
dev.off()
 
 wrap.cats <- c("access",
                "awareness",
                "intention")
 
dat %>%  filter(cat %in% wrap.cats) %>% ggplot(aes(x = val, y = excess)) + geom_bar(stat = "identity") + ylab("Excess NF Mass %") +
  xlab("Likert rating (Higher rating = more automatic)") + facet_wrap(~ cat, scales = "free", labeller = fun_labeller)

 dev.copy(jpeg,'Output/Graphs/Experiments/Automaticity/excess mass.jpg')
dev.off()

choice.long2 %>%  filter(cat %in% wrap.cats) %>%
  ggplot(aes(x=val)) + stat_ecdf(aes(colour=target)) + facet_wrap(~ cat)

'%!in%' <- function(x,y)!('%in%'(x,y))

choice.long2 %>% filter(cat %!in% c("speed", "careful", "gut")) %>%
  ggplot(aes(x=val)) + stat_ecdf(aes(colour=target)) + facet_wrap(~ cat, scales = "free")

dev.copy(jpeg,'Output/Graphs/Experiments/Automaticity/cdf for each measure.jpg')
dev.off()

```

```{r continuous measures - median split}


label_names <- list(
  'cogeffort'="(Cog Effort) \nLack of forethought",
  'control'="(Control) \nInability to resist action urge",
  'efficiency'="(Efficiency) \nLow cognitive effort",
  'influence'="(Influence) \nInfleunced by situation",
  'access' = "(Access) \nDifficult to explain why",
  'awareness' = "(Unwareness) \nUnaware what affected decision",
  'intention' = "(Intention) \nInability to resist considering"
)

fun_labeller <- function(variable,value){
  return(label_names[value])
}

dat <- m.long %>% group_by(cat) %>% 
       mutate(cat_median = median(val, na.rm = T)) %>% 
       ungroup

dat$above_median <- dat$val >= dat$cat_median

'%!in%' <- function(x,y)!('%in%'(x,y))

l.dat <- dat %>% filter(cat %!in% c("speed", "careful", "gut")) %>% 
  group_by(cat, target)  %>% 
  dplyr::summarise(percent.above = mean(above_median, na.rm = T), sd = sd(above_median, na.rm = T), n= n())

l.dat$ci <- (l.dat$sd/sqrt(l.dat$n))*1.96
pd <- position_dodge(0.1)

 ggplot(l.dat, aes(x = target, y = percent.above)) +
  geom_bar(stat="identity", position=pd) +
   geom_errorbar(aes(ymin=percent.above-ci, ymax=percent.above+ci), colour="black", width=.1, position=pd) +
   theme(axis.text.x = element_text(angle = 45, hjust = 1)) + xlab("Measure") +
   ylab("Percent of responses above median") + xlab("") + labs(title = "Median splits of automaticity Ratings by Target") + facet_wrap(~ cat, scales = "free", labeller = fun_labeller)
 
  dev.copy(jpeg,'Output/Graphs/Experiments/Automaticity/bar chart pct above median.jpg')
dev.off()
  
```

```{r binary measures}

label_names <- list(
  'careful'="(Careful) Not a careful consideration",
  'gut'="(Gut) Based on gut reaction"
)

fun_labeller <- function(variable,value){
  return(label_names[value])
}

l.dat <- m.long %>% filter(cat %in% c("careful", "gut")) %>% 
  group_by(cat, target)  %>% 
  dplyr::summarise(mean = mean(val, na.rm = T), sd = sd(val, na.rm = T), n= n())

l.dat$ci <- (l.dat$sd/sqrt(l.dat$n))*1.96
pd <- position_dodge(0.1)

l.dat$target <- with(l.dat, ifelse(target == "nf", "Newsfeed", "PYMK"))

 ggplot(l.dat, aes(x = target, y = mean, fill = target)) +
  geom_bar(stat="identity", position=pd) +
   geom_errorbar(aes(ymin=mean-ci, ymax=mean+ci), colour="black", width=.1, position=pd) +
   theme(axis.text.x = element_text(angle = 45, hjust = 1)) + xlab("Measure") +
   ylab("Percent giving more automatic answer") + xlab("") + labs(title = "Rate of Affirming Automaticity by Target") + facet_wrap(~ cat, scales = "free", labeller = fun_labeller) +
   scale_fill_manual(values=colors)
 
  dev.copy(jpeg,'Output/Graphs/Experiments/Automaticity/bar chart binary measures.jpg')
dev.off()

```

```{r}

m.long %>% filter(cat == "speed") %>%
  ggplot(aes(x = val, fill = target)) +
  geom_histogram(aes(fill = target, y = ..density..), position = "dodge") +
  labs(title = "Histogram of reported time to decide")

 dev.copy(jpeg,'Output/Graphs/Experiments/Automaticity/speed histogram.jpg')
dev.off()

m.long %>% filter(cat == "speed") %>%
  mutate(Seconds = ifelse(val > 60, 60, val)) %>% 
  ggplot(aes(x = Seconds, fill = target)) +
  geom_histogram(aes(fill = target, y = ..density..), position = "dodge") + labs(title = "Histogram of reported time to decide",
                                                                subtitle = "bin over 60")

 dev.copy(jpeg,'Output/Graphs/Experiments/Automaticity/speed histogram - bin over 60.jpg')
dev.off()

m.long %>% filter(cat == "speed") %>%
  mutate(Seconds = ifelse(val > 30, 30, val)) %>% 
  ggplot(aes(x = Seconds, fill = target)) +
  geom_histogram(aes(fill = target, y = ..density..), position = "dodge") + labs(title = "Histogram of reported time to decide",
                                                                subtitle = "bin over 30")

 dev.copy(jpeg,'Output/Graphs/Experiments/Automaticity/speed histogram - bin over 30.jpg')
dev.off()

m.long %>% filter(cat == "speed") %>% 
  mutate(Seconds = ifelse(val > 60, 60, val)) %>%
ggplot(aes(x=Seconds)) + stat_ecdf(aes(colour=target)) +
   scale_color_manual(values=colors) + ylab("Cumulative Density")

dev.copy(jpeg,'Output/Graphs/Experiments/Automaticity/speed cdf - bin over 60.jpg')
dev.off()

```
